Organized the mod API
