+++ src/main/java/T145/metalchests/client/render/blocks/RenderHungryMetalChest.java
@@ -15,7 +15,7 @@
  ******************************************************************************/
 package T145.metalchests.client.render.blocks;
 
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
+import T145.metalchests.api.immutable.ChestType;
 import T145.metalchests.core.MetalChests;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.relauncher.Side;
