+++ src/main/java/T145/metalchests/client/render/blocks/RenderMetalChest.java
@@ -17,7 +17,7 @@ package T145.metalchests.client.render.blocks;
 
 import org.lwjgl.opengl.GL11;
 
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
+import T145.metalchests.api.immutable.ChestType;
 import T145.metalchests.core.MetalChests;
 import T145.metalchests.tiles.TileMetalChest;
 import net.minecraft.client.model.ModelChest;
