+++ src/main/java/T145/metalchests/client/render/blocks/RenderSortingMetalChest.java
@@ -20,7 +20,7 @@ import javax.annotation.Nullable;
 import org.lwjgl.opengl.GL11;
 
 import T145.metalchests.api.BlocksMetalChests;
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
+import T145.metalchests.api.immutable.ChestType;
 import T145.metalchests.core.MetalChests;
 import T145.metalchests.tiles.TileSortingMetalChest;
 import net.blay09.mods.refinedrelocation.RefinedRelocationConfig;
