+++ src/main/java/T145/metalchests/core/ModLoader.java
@@ -22,8 +22,8 @@ import java.util.Map;
 import T145.metalchests.api.BlocksMetalChests;
 import T145.metalchests.api.ItemsMetalChests;
 import T145.metalchests.api.ModSupport;
+import T145.metalchests.api.immutable.ChestType;
 import T145.metalchests.blocks.BlockMetalChest;
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
 import T145.metalchests.client.render.blocks.RenderMetalChest;
 import T145.metalchests.client.render.entities.RenderMinecartMetalChest;
 import T145.metalchests.config.ModConfig;
