+++ src/main/java/T145/metalchests/core/modules/ModuleThaumcraft.java
@@ -21,8 +21,8 @@ import java.util.Map;
 import T145.metalchests.api.BlocksMetalChests;
 import T145.metalchests.api.ItemsMetalChests;
 import T145.metalchests.api.ModSupport;
+import T145.metalchests.api.immutable.ChestType;
 import T145.metalchests.blocks.BlockHungryMetalChest;
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
 import T145.metalchests.client.render.blocks.RenderHungryMetalChest;
 import T145.metalchests.core.MetalChests;
 import T145.metalchests.core.ModLoader;
