+++ src/main/java/T145/metalchests/lib/DataSerializers.java
@@ -17,7 +17,7 @@ package T145.metalchests.lib;
 
 import java.io.IOException;
 
-import T145.metalchests.blocks.BlockMetalChest.ChestType;
+import T145.metalchests.api.immutable.ChestType;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.network.datasync.DataParameter;
 import net.minecraft.network.datasync.DataSerializer;
